# api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockNew API",
    "description": "Local async API surface over the StockNew quant platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/backtests": {
      "post": {
        "tags": [
          "backtests"
        ],
        "summary": "Trigger Backtest",
        "operationId": "trigger_backtest_api_backtests_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "backtests"
        ],
        "summary": "List Backtests",
        "operationId": "list_backtests_api_backtests_get",
        "parameters": [
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "exclusiveMinimum": 0,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BacktestSummary"
                  },
                  "title": "Response List Backtests Api Backtests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/backtests/{run_id}": {
      "get": {
        "tags": [
          "backtests"
        ],
        "summary": "Get Backtest",
        "operationId": "get_backtest_api_backtests__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics": {
      "post": {
        "tags": [
          "diagnostics"
        ],
        "summary": "Trigger Diagnostic",
        "operationId": "trigger_diagnostic_api_diagnostics_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiagnosticRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "List Diagnostics",
        "operationId": "list_diagnostics_api_diagnostics_get",
        "parameters": [
          {
            "name": "factor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Factor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "exclusiveMinimum": 0,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiagnosticSummary"
                  },
                  "title": "Response List Diagnostics Api Diagnostics Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics/{run_id}": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "summary": "Get Diagnostic",
        "operationId": "get_diagnostic_api_diagnostics__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestRequest": {
        "properties": {
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "default": "swing_trading"
          },
          "universe": {
            "type": "string",
            "enum": [
              "watchlist",
              "portfolio",
              "themes",
              "tickers"
            ],
            "title": "Universe",
            "default": "themes"
          },
          "tickers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tickers",
            "description": "Required when universe='tickers'. Ignored otherwise."
          },
          "years": {
            "type": "number",
            "maximum": 20.0,
            "exclusiveMinimum": 0.0,
            "title": "Years",
            "default": 3.0
          },
          "min_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Score"
          },
          "max_positions": {
            "type": "integer",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Positions",
            "default": 20
          },
          "position_pct": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Position Pct",
            "default": 0.1
          },
          "cash": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Cash",
            "default": 10000.0
          },
          "hold_days": {
            "type": "integer",
            "maximum": 400.0,
            "exclusiveMinimum": 0.0,
            "title": "Hold Days",
            "default": 90
          },
          "earnings_blackout": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Earnings Blackout",
            "default": 3
          },
          "accept_lookahead": {
            "type": "boolean",
            "title": "Accept Lookahead",
            "default": false
          },
          "oos_split": {
            "type": "number",
            "maximum": 0.6,
            "minimum": 0.0,
            "title": "Oos Split",
            "default": 0.3
          },
          "bootstrap_resamples": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Bootstrap Resamples",
            "default": 0
          },
          "commission": {
            "type": "number",
            "minimum": 0.0,
            "title": "Commission",
            "default": 0.0
          },
          "slippage_bps": {
            "type": "number",
            "minimum": 0.0,
            "title": "Slippage Bps",
            "default": 5.0
          },
          "regulatory_bps": {
            "type": "number",
            "minimum": 0.0,
            "title": "Regulatory Bps",
            "default": 3.0
          },
          "vol_target_risk": {
            "type": "number",
            "maximum": 0.1,
            "minimum": 0.0,
            "title": "Vol Target Risk",
            "default": 0.0
          },
          "fresh": {
            "type": "boolean",
            "title": "Fresh",
            "default": false
          }
        },
        "type": "object",
        "title": "BacktestRequest"
      },
      "BacktestResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "id",
          "strategy",
          "window_start",
          "window_end",
          "result"
        ],
        "title": "BacktestResponse"
      },
      "BacktestSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "strategy": {
            "type": "string",
            "title": "Strategy"
          },
          "universe_label": {
            "type": "string",
            "title": "Universe Label"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "n_trades": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Trades"
          },
          "oos_sharpe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oos Sharpe"
          },
          "oos_total_return_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oos Total Return Pct"
          },
          "oos_max_drawdown_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oos Max Drawdown Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "strategy",
          "universe_label",
          "window_start",
          "window_end",
          "created_at"
        ],
        "title": "BacktestSummary"
      },
      "DiagnosticRequest": {
        "properties": {
          "strategy": {
            "type": "string",
            "title": "Strategy",
            "default": "swing_trading"
          },
          "universe": {
            "type": "string",
            "enum": [
              "watchlist",
              "portfolio",
              "themes",
              "tickers"
            ],
            "title": "Universe",
            "default": "themes"
          },
          "tickers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tickers"
          },
          "years": {
            "type": "number",
            "maximum": 10.0,
            "exclusiveMinimum": 0.0,
            "title": "Years",
            "default": 2.0
          },
          "factor": {
            "type": "string",
            "enum": [
              "composite",
              "technical",
              "fundamental",
              "pattern",
              "statistical",
              "trend",
              "alpha158"
            ],
            "title": "Factor",
            "default": "composite"
          },
          "quantiles": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 2.0,
            "title": "Quantiles",
            "default": 5
          },
          "periods": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Periods",
            "description": "Forward-return horizons in trading days."
          },
          "accept_lookahead": {
            "type": "boolean",
            "title": "Accept Lookahead",
            "default": false
          },
          "fresh": {
            "type": "boolean",
            "title": "Fresh",
            "default": false
          }
        },
        "type": "object",
        "title": "DiagnosticRequest"
      },
      "DiagnosticResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "factor": {
            "type": "string",
            "title": "Factor"
          },
          "universe_label": {
            "type": "string",
            "title": "Universe Label"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "quantiles": {
            "type": "integer",
            "title": "Quantiles"
          },
          "n_observations": {
            "type": "integer",
            "title": "N Observations"
          },
          "ic_mean": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ic Mean"
          },
          "ic_std": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ic Std"
          },
          "ic_ir": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Ic Ir"
          },
          "top_minus_bottom_pct": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Top Minus Bottom Pct"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "factor",
          "universe_label",
          "window_start",
          "window_end",
          "created_at",
          "quantiles",
          "n_observations",
          "ic_mean",
          "ic_std",
          "ic_ir",
          "top_minus_bottom_pct"
        ],
        "title": "DiagnosticResponse"
      },
      "DiagnosticSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "factor": {
            "type": "string",
            "title": "Factor"
          },
          "universe_label": {
            "type": "string",
            "title": "Universe Label"
          },
          "window_start": {
            "type": "string",
            "format": "date-time",
            "title": "Window Start"
          },
          "window_end": {
            "type": "string",
            "format": "date-time",
            "title": "Window End"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "n_observations": {
            "type": "integer",
            "title": "N Observations"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "factor",
          "universe_label",
          "window_start",
          "window_end",
          "created_at",
          "n_observations"
        ],
        "title": "DiagnosticSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}